// frontend/src/apisv2/run/models/V2Run.ts
export interface V2Run {
  /**
   * Input. ID of the parent experiment.
   * The default experiment ID will be used if this is not specified.
   * @type {string}
   * @memberof V2Run
   */
  experiment_id?: string;
  /**
   * Output. Unique run ID. Generated by API server.
   * @type {string}
   * @memberof V2Run
   */
  run_id?: string;
  /**
   * Required input. Name provided by user,
   * or auto generated if run is created by a recurring run.
   * @type {string}
   * @memberof V2Run
   */
  display_name?: string;
  /**
   *
   * @type {V2RunStorageState}
   * @memberof V2Run
   */
  storage_state?: V2RunStorageState;
  /**
   * Optional input. Short description of the run.
   * @type {string}
   * @memberof V2Run
   */
  description?: string;
  /**
   * This field is Deprecated. The pipeline version id is under pipeline_version_reference for v2.
   * @type {string}
   * @memberof V2Run
   */
  pipeline_version_id?: string;
  /**
   * Pipeline spec.
   * @type {object}
   * @memberof V2Run
   */
  pipeline_spec?: object;
  /**
   *
   * @type {V2PipelineVersionReference}
   * @memberof V2Run
   */
  pipeline_version_reference?: V2PipelineVersionReference;
  /**
   *
   * @type {V2RuntimeConfig}
   * @memberof V2Run
   */
  runtime_config?: V2RuntimeConfig;
  /**
   * Optional input. Specifies which kubernetes service account is used.
   * @type {string}
   * @memberof V2Run
   */
  service_account?: string;
  /**
   * Output. Creation time of the run.
   * @type {Date}
   * @memberof V2Run
   */
  created_at?: Date;
  /**
   * Output. When this run is scheduled to start. This could be different from
   * created_at. For example, if a run is from a backfilling job that was supposed
   * to run 2 month ago, the created_at will be 2 month behind scheduled_at.
   * @type {Date}
   * @memberof V2Run
   */
  scheduled_at?: Date;
  /**
   * Output. Completion of the run.
   * @type {Date}
   * @memberof V2Run
   */
  finished_at?: Date;
  /**
   *
   * @type {V2RuntimeState}
   * @memberof V2Run
   */
  state?: V2RuntimeState;
  /**
   *
   * @type {GoogleRpcStatus}
   * @memberof V2Run
   */
  error?: GoogleRpcStatus;
  /**
   *
   * @type {V2RunDetails}
   * @memberof V2Run
   */
  run_details?: V2RunDetails;
  /**
   * ID of the recurring run that triggered this run.
   * @type {string}
   * @memberof V2Run
   */
  recurring_run_id?: string;
  /**
   * Output. A sequence of run statuses. This field keeps a record
   * of state transitions.
   * @type {Array<V2RuntimeStatus>}
   * @memberof V2Run
   */
  state_history?: Array<V2RuntimeStatus>;
  /**
   * Optional input. Plugin-specific inputs provided by the user at run creation.
   * Each key is a plugin name (e.g., "mlflow") and the value is arbitrary JSON config.
   * @type {{ [key: string]: object; }}
   * @memberof V2Run
   */
  plugins_input?: { [key: string]: object };
  /**
   * Output. Plugin-specific outputs populated by backend components.
   * Each key is a plugin name and the value contains the plugin's output entries and state.
   * @type {{ [key: string]: V2PluginOutput; }}
   * @memberof V2Run
   */
  plugins_output?: { [key: string]: V2PluginOutput };
  /**
   *
   * @type {V2PipelineVersionReference}
   * @memberof V2Run
   */
  pipeline_reference?: V2PipelineVersionReference;
  /**
   *
   * @type {number}
   * @memberof V2Run
   */
  task_count?: number;
  /**
   *
   * @type {Array<V2PipelineTask>}
   * @memberof V2Run
   */
  tasks?: Array<V2PipelineTask>;
}